requirements.txt
aiohttp==3.8.4
autogen==0.7.2
beautifulsoup4==4.12.3
dlib==19.24.6
fake_useragent==1.1.3
Flask==3.1.0
Flask_Cors==5.0.0
imutils==0.5.4
langchain==0.3.15
langchain_community==0.3.15
langchain_core==0.3.31
langchain_huggingface==0.1.2
numpy==1.26.4
openai==1.60.1
opencv_python==4.10.0.84
python-dotenv==1.0.1
Requests==2.32.3
tavily_python==0.5.0
tensorflow==2.18.0
tensorflow_intel==2.18.0
